requirements.txt
fastapi[all]
requests
asyncio
pycryptodome